Revert "Create a util for Crashlytics tasks, and move race into it (#6158)"
This reverts commit 9ed73c407bee8bde1e7c6e611a68f1639374bf9c.
